Brief-5-Challenges
Timer added to GameManager that starts at 30 seconds and counts down. When a target is hit, 5 seconds is added to the timer.

Game difficulty changes when the 1-3 keys are pressed. Target sizes will change according to the difficulty.

Weapon select system has been migrated to the 'Q' key which will cycle through the projectiles.