Week-6-Brief-Requirements-And-Challenges
Created canvas with text objects for score, targets left, and time.

Created UIManager as a Singleton that displays the the score from the GameManager and the amount of targets left in the scene.

Timer was added in last commit.

Created a title screen that allows us to start the game and quit from the game.

Added visual timer to the UI

Added options screen with a difficulty drop down in the title screen.

Current Weapon name is displayed in the UI.

Targets now spawn automatically.

Game Over screen added.